fix(automate): multiregion event listeners
